Rename and organize query clients
