--- nibiru/clients/__init__.py
@@ -1,4 +1,0 @@
-from nibiru.clients.dex import DexQueryClient  # noqa
-from nibiru.clients.perp import PerpQueryClient  # noqa
-from nibiru.clients.pricefeed import PricefeedQueryClient  # noqa
-from nibiru.clients.vpool import VpoolQueryClient  # noqa
