+++ nibiru/clients/query/__init__.py
@@ -1,0 +1,4 @@
+from .dex import DexQueryClient  # noqa
+from .perp import PerpQueryClient  # noqa
+from .pricefeed import PricefeedQueryClient  # noqa
+from .vpool import VpoolQueryClient  # noqa
