+++ nibiru/clients/query/perp.py
@@ -3,8 +3,9 @@ from grpc import Channel
 from nibiru_proto.proto.perp.v1 import query_pb2 as perp_type
 from nibiru_proto.proto.perp.v1 import query_pb2_grpc as perp_query
 
-from nibiru.clients.util import deserialize
 from nibiru.utils import from_sdk_dec
+
+from .util import deserialize
 
 
 class PerpQueryClient:
