+++ nibiru/clients/query/vpool.py
@@ -3,8 +3,9 @@ from nibiru_proto.proto.vpool.v1 import query_pb2 as vpool_type
 from nibiru_proto.proto.vpool.v1 import query_pb2_grpc as vpool_query
 from nibiru_proto.proto.vpool.v1.vpool_pb2 import Direction as pbDirection
 
-from nibiru.clients.util import deserialize
 from nibiru.common import Direction
+
+from .util import deserialize
 
 
 class VpoolQueryClient:
